requirements.txt
django
gunicorn
django-filter
psycopg2-binary
django-cors-headers
djangorestframework


